requirements.txt
grpcio-tools==1.42.0
jinja2==3.1.2
netifaces==0.11.0
markupsafe==2.1.3
requests==